redactor(errors): consolidate error handling
